refactor(routes): extract /channels into its own module (god-file decomposition)
Channel routes moved verbatim from mycelium.js into server/routes/channels.js
(handler bodies byte-identical, audit auth/guardrail calls preserved), wired via
registerChannelRoutes(router, deps). mycelium.js -189. Route contract unchanged
(route-manifest --check: 284 routes), vitest 265/265. By Lucy (GLM-5.2), byte-reviewed by m5Max.
